split out compiler_llvm.py
